requirements.txt
Pillow==7.2.0
pkg-resources==0.0.0
pyTelegramBotAPI==3.7.2
requests==2.7.0
six==1.15.0
